+++ apps/notifications/src/common/config/event-key-registry.config.ts
@@ -1,5 +1,5 @@
-export const eventKeyRegistry = Object.freeze({
+export const eventKeyRegistry = {
   blockCreated: "blockchain.v1.block.created",
   eventCloseDeployment: "akash.v1.deployment.deployment-closed",
   createNotification: "notifications.v1.notification.create"
-});
+} as const;