+++ apps/notifications/CHANGELOG.md
@@ -1,5 +1,19 @@
 
 
+## [1.21.0](https://github.com/akash-network/console/compare/notifications/v1.20.3...notifications/v1.21.0) (2025-08-14)
+
+
+### Features
+
+* enables sentry sourcemaps in deploy-web ([#1800](https://github.com/akash-network/console/issues/1800)) ([f7c83bf](https://github.com/akash-network/console/commit/f7c83bf749199d17e9d9b8cb7c2f7a3413a59887))
+
+## [1.20.3](https://github.com/akash-network/console/compare/notifications/v1.20.2...notifications/v1.20.3) (2025-08-07)
+
+
+### Code Refactoring
+
+* archive completed jobs in 24h ([#1783](https://github.com/akash-network/console/issues/1783)) ([219e0a3](https://github.com/akash-network/console/commit/219e0a37a0073feaba78d35af063e03f5dea6c65))
+
 ## [1.20.2](https://github.com/akash-network/console/compare/notifications/v1.20.1...notifications/v1.20.2) (2025-08-06)