+++ apps/provider-console/CHANGELOG.md
@@ -1,5 +1,19 @@
 
 
+## [1.20.0](https://github.com/akash-network/console/compare/provider-console/v1.19.1...provider-console/v1.20.0) (2025-08-14)
+
+
+### Features
+
+* enables sentry sourcemaps in deploy-web ([#1800](https://github.com/akash-network/console/issues/1800)) ([f7c83bf](https://github.com/akash-network/console/commit/f7c83bf749199d17e9d9b8cb7c2f7a3413a59887))
+
+## [1.19.1](https://github.com/akash-network/console/compare/provider-console/v1.19.0...provider-console/v1.19.1) (2025-08-07)
+
+
+### Bug Fixes
+
+* **provider:** added loading for api-key get call ([#1792](https://github.com/akash-network/console/issues/1792)) ([9552aa2](https://github.com/akash-network/console/commit/9552aa2c711c0613b2efd92d4db2aae75abf328f))
+
 ## [1.19.0](https://github.com/akash-network/console/compare/provider-console/v1.18.0...provider-console/v1.19.0) (2025-07-31)